// management/src/routes/comments/__generated__/CommentsTableFragment.graphql.ts
export type CommentStatus = "approved" | "pending" | "spam" | "trash" | "%future added value";
import { FragmentRefs } from "relay-runtime";
export type CommentsTableFragment$data = {
  readonly comments: {
    readonly edges: ReadonlyArray<{
      readonly cursor: string;
      readonly node: {
        readonly authorName: string | null | undefined;
        readonly content: string;
        readonly id: string;
        readonly status: CommentStatus;
        readonly submittedAt: any;
      } | null | undefined;
    }>;
    readonly pageInfo: {
      readonly endCursor: string | null | undefined;
      readonly hasNextPage: boolean;
    };
    readonly totalCount: number;
  };
  readonly " $fragmentType": "CommentsTableFragment";
};
export type CommentsTableFragment$key = {
  readonly " $data"?: CommentsTableFragment$data;
  readonly " $fragmentSpreads": FragmentRefs<"CommentsTableFragment">;
};